Refactor Yacht component
