+++ src/components/Navbar/Navbar.module.css
@@ -2,26 +2,22 @@
   position: fixed;
   width: 20%;
   height: 100vh;
-  border-right: 0.5px solid #efefef;
+  border-right: 0.5px solid #EFEFEF;
   z-index: 10000;
 }
-
 .nav {
   position: relative;
   top: 50%;
   transform: translateY(-70%);
   font-weight: bold;
 }
-
 .nav ul {
   list-style: none;
   padding-left: 10%;
 }
-
 .nav ul li {
   position: relative;
 }
-
 li a {
   display: block;
   text-decoration: none;
@@ -30,11 +26,9 @@ li a {
   transition: all 0.3s ease-in-out;
   padding: 1em;
 }
-
 .logo {
   width: 100%;
 }
-
 .logoutButton {
   width: 100%;
   padding: 0.6em 0.6em 0.6em 1em;
@@ -44,11 +38,9 @@ li a {
   border: none;
   text-align: left;
 }
-
 .hamburgerMenu {
   display: none;
 }
-
 @media only screen and (max-width: 800px) {
   .hamburgerMenu {
     display: block;
@@ -59,29 +51,24 @@ li a {
     left: 10px;
     z-index: 10001;
   }
-
   .hamburgerButton,
   .hamburgerImage {
     width: 100%;
     height: 100%;
   }
-
   .hamburgerButton {
     border: none;
     background: transparent;
   }
-
   .sidebar {
     display: none;
     min-width: 100vw;
     min-height: 100vh;
   }
-
   .open {
     display: block;
     background: #fff;
   }
-
   .nav {
     transform: none;
     font-weight: bold;
@@ -95,20 +82,16 @@ li a {
     flex-direction: column;
     justify-content: center;
   }
-
   .logo {
     width: 40%;
     margin: 0 auto;
   }
-
   .nav ul {
     padding-left: 0;
   }
-
   .nav ul li {
     margin-bottom: 1rem;
   }
-
   .logoutButton {
     width: 25%;
     text-align: center;
